+++ DotSerial.Tests/Core/XML/XMLSerialTests.cs
@@ -156,6 +156,21 @@ namespace DotSerial.Tests.Core.XML
         }
 
         [Fact]
+        public void CreateSerializedObject_ParsableClass()
+        {
+            // Arrange
+            var tmp = ParsableClass.CreateTestDefault();
+
+            // Act
+            var xmlDocument = DotSerialXML.Serialize(tmp);
+            var result = DotSerialXML.Deserialize<ParsableClass>(xmlDocument);
+
+            // Assert
+            Assert.NotNull(result);
+            EqualCheck.AssertClassEqual(tmp, result);
+        }
+
+        [Fact]
         public void CreateSerializedObject_PrimitiveClass()
         {
             // Arrange